Migrate Solution file to XML Based Solution File, fixed typo and modernised some of the code
