+++ Shared/GameState.cs
@@ -1,10 +1,9 @@
-﻿namespace WordleBlazorApp.Shared
+﻿namespace WordleBlazorApp.Shared;
+
+public class GameState
 {
-    public class GameState
-    {
-        public bool IsGameComplete { get; set; }
-        public bool HasPlayerWonGame { get; set; }
-        public int NumberOfGuesses { get; set; }
-        public GuessResult GuessResult { get; set; }
-    }
+    public bool IsGameComplete { get; set; }
+    public bool HasPlayerWonGame { get; set; }
+    public int NumberOfGuesses { get; set; }
+    public GuessResult GuessResult { get; set; } = null!;
 }
