+++ WordleBlazorMaui/WordleBlazorMaui.sln
@@ -6,7 +6,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WordleBlazorMaui", "WordleB
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Shared", "..\Shared\Shared.csproj", "{F3BC9C21-554D-4E5D-B5E4-721A3089FA46}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WorldleGameEngine", "..\WorldleGameEngine\WorldleGameEngine.csproj", "{8462C78E-D46D-49C6-8484-4B24D58C458B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WorldleGameEngine", "..\WordleGameEngine\WordleGameEngine.csproj", "{8462C78E-D46D-49C6-8484-4B24D58C458B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
