+++ WordleBlazorMaui/WordleBlazorMaui/App.xaml.cs
@@ -6,6 +6,6 @@ public partial class App : Application
 	{
 		InitializeComponent();
 
-		MainPage = new MainPage();
+		//MainPage = new MainPage();
 	}
 }
