+++ WordleBlazorMaui/WordleBlazorMaui/WordleBlazorMaui.csproj
@@ -64,7 +64,7 @@
 
 	<ItemGroup>
 	  <ProjectReference Include="..\..\Shared\Shared.csproj" />
-	  <ProjectReference Include="..\..\WorldleGameEngine\WorldleGameEngine.csproj" />
+	  <ProjectReference Include="..\..\WordleGameEngine\WordleGameEngine.csproj" />
 	</ItemGroup>
 
 </Project>
