+++ WordleBlazorMaui/WordleBlazorMaui/_Imports.razor
@@ -4,8 +4,8 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
-@using WorldleGameEngine.Interfaces
-@using WorldleGameEngine
+@using WordleGameEngine.Interfaces
+@using WordleGameEngine
 @using WordleBlazorApp.Shared
 @using WordleBlazorMaui
 @using WordleBlazorMaui.Shared
