+++ WordleBlazorMaui/WordleBlazorMaui/appsettings.json
@@ -1,5 +1,5 @@
 {
   "PossibleWordles": [
-
+    "tears", "blame", "crops", "straw", "slope", "grace", "flame"
   ]
 }
