--- WordleBlazorServerApp/WordleBlazorServerApp.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.32112.339
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WordleBlazorServerApp", "WordleBlazorServerApp\WordleBlazorServerApp.csproj", "{13B213CC-E25A-450B-94E9-16E8B32AF03C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Shared", "..\Shared\Shared.csproj", "{E37ADACB-481E-492C-9122-84FF0B40C735}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WorldleGameEngine", "..\WorldleGameEngine\WorldleGameEngine.csproj", "{364554FC-ED92-4CBD-9CB9-2B2C7D58DBD1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{13B213CC-E25A-450B-94E9-16E8B32AF03C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{13B213CC-E25A-450B-94E9-16E8B32AF03C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{13B213CC-E25A-450B-94E9-16E8B32AF03C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{13B213CC-E25A-450B-94E9-16E8B32AF03C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E37ADACB-481E-492C-9122-84FF0B40C735}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E37ADACB-481E-492C-9122-84FF0B40C735}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E37ADACB-481E-492C-9122-84FF0B40C735}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E37ADACB-481E-492C-9122-84FF0B40C735}.Release|Any CPU.Build.0 = Release|Any CPU
-		{364554FC-ED92-4CBD-9CB9-2B2C7D58DBD1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{364554FC-ED92-4CBD-9CB9-2B2C7D58DBD1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{364554FC-ED92-4CBD-9CB9-2B2C7D58DBD1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{364554FC-ED92-4CBD-9CB9-2B2C7D58DBD1}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C06305BF-92BD-4C56-80AE-2F672CB46B66}
-	EndGlobalSection
-EndGlobal
