+++ WordleBlazorServerApp/WordleBlazorServerApp/_Imports.razor
@@ -5,8 +5,8 @@
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.Web.Virtualization
-@using WorldleGameEngine.Interfaces
-@using WorldleGameEngine
+@using WordleGameEngine.Interfaces
+@using WordleGameEngine
 @using WordleBlazorApp.Shared
 @using WordleBlazorServerApp
 @using WordleBlazorServerApp.Shared
