+++ WordleBlazorWASMApp/WordleBlazorWASMApp/WordleBlazorWASMApp.csproj
@@ -14,7 +14,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\Shared\Shared.csproj" />
-    <ProjectReference Include="..\..\WorldleGameEngine\WorldleGameEngine.csproj" />
+    <ProjectReference Include="..\..\WordleGameEngine\WordleGameEngine.csproj" />
   </ItemGroup>
 
   <ItemGroup>
