+++ WordleBlazorWASMApp/WordleBlazorWASMApp/_Imports.razor
@@ -6,8 +6,8 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
-@using WorldleGameEngine.Interfaces
-@using WorldleGameEngine
+@using WordleGameEngine.Interfaces
+@using WordleGameEngine
 @using WordleBlazorApp.Shared
 @using WordleBlazorWASMApp
 @using WordleBlazorWASMApp.Shared
