+++ WordleGameConsoleApp/WordleGameConsoleApp/WordleGameConsoleApp.csproj
@@ -1,0 +1,21 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net10.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+	<PackageReference Include="Microsoft.Extensions.Configuration" Version="10.0.2" />
+	<PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="10.0.2" />
+	<PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="10.0.2" />
+	<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="10.0.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\WordleGameEngine\WordleGameEngine.csproj" />
+  </ItemGroup>
+
+</Project>
