+++ WordleGameEngine/Interfaces/IGameEngine.cs
@@ -1,0 +1,11 @@
+﻿using WordleBlazorApp.Shared;
+
+namespace WordleGameEngine.Interfaces;
+
+public interface IGameEngine
+{
+    GameState NewGame();
+    GameState EnterGuess(string guess);
+    int GetNumberOfGuesses();
+    GameGrid GetGameGrid();
+}
