+++ WordleGameEngine/Interfaces/IWordleGenerator.cs
@@ -1,0 +1,6 @@
+﻿namespace WordleGameEngine.Interfaces;
+
+public interface IWordleGenerator
+{
+    string GenerateSelectedWordle();
+}
