+++ WordleGameEngine/WordleGenerator.cs
@@ -1,0 +1,33 @@
+﻿using WordleGameEngine.Interfaces;
+
+namespace WordleGameEngine;
+
+public class WordleGenerator : IWordleGenerator
+{
+    private readonly List<string> _possibleWordles;
+
+    public WordleGenerator(List<string> possibleWordles)
+    {
+        _possibleWordles = possibleWordles;
+    }
+
+    public string GenerateSelectedWordle()
+    {
+        var possibleWordles = GetPossibleWordles();
+
+        var random = new Random();
+        int wordlePostion = random.Next(0, possibleWordles.Count);
+
+        return possibleWordles[wordlePostion];
+    }
+
+    private List<string> GetPossibleWordles()
+    {
+        return _possibleWordles
+                    .GroupBy(x => x)
+                    .Select(x => x.Key.ToLower())
+                    .ToList()
+                    .OrderBy(x => x)
+                    .ToList();
+    }
+}
