+++ WordleGameEngineUnitTests/Usings.cs
@@ -1,4 +1,4 @@
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using Moq;
-global using WorldleGameEngine;
-global using WorldleGameEngine.Interfaces;
+global using WordleGameEngine;
+global using WordleGameEngine.Interfaces;
