--- WorldleGameEngine/Interfaces/IGameEngine.cs
@@ -1,12 +1,0 @@
-﻿using WordleBlazorApp.Shared;
-
-namespace WorldleGameEngine.Interfaces
-{
-    public interface IGameEngine
-    {
-        GameState NewGame();
-        GameState EnterGuess(string guess);
-        int GetNumberOfGuesses();
-        GameGrid GetGameGrid();
-    }
-}
