--- WorldleGameEngine/Interfaces/IWordleGenerator.cs
@@ -1,7 +1,0 @@
-﻿namespace WorldleGameEngine.Interfaces
-{
-    public interface IWordleGenerator
-    {
-        string GenerateSelectedWordle();
-    }
-}
